backend/app/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pandas
numpy
python-multipart
